Revert "add more env + fix auth + struct"
