--- .env.example
@@ -1,2 +1,0 @@
-VITE_APP_RECAPTCHA_SITE_KEY=
-BASE_URL=
