+++ package.json
@@ -12,13 +12,11 @@
   "dependencies": {
     "@ant-design/icons": "^6.0.0",
     "@tailwindcss/vite": "^4.1.10",
-    "@tanstack/react-query": "^5.81.2",
     "antd": "^5.26.2",
     "axios": "^1.10.0",
     "js-cookie": "^3.0.5",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-google-recaptcha": "^3.1.0",
     "react-router-dom": "^7.6.2",
     "zustand": "^5.0.5"
   },
