--- src/components/layouts/LandingLayout.jsx
@@ -1,11 +1,0 @@
-const LandingLayout = ({ children }) => {
-  return (
-    <>
-      <div>header</div>
-      <main>{children}</main>
-      <div>footer</div>
-    </>
-  );
-};
-
-export default LandingLayout;
