+++ src/configs/axiosClient.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const axiosClient = axios.create({
-  baseURL: import.meta.env.BASE_URL || 'http://localhost:3000',
+  baseURL: '',
   headers: {
     'Content-Type': 'application/json',
   },
