--- src/constants/index.js
@@ -1,11 +1,0 @@
-const PATH_NAME = {
-  HOME: '/',
-  AUTH: '/auth',
-  POST: '/post',
-  POST_DETAIL: '/post/:id',
-  NOT_FOUND: '*',
-  MEMBER: '/member',
-  ADMIN: '/admin',
-};
-
-export { PATH_NAME };
