--- src/pages/notfound/index.jsx
@@ -1,9 +1,0 @@
-const NotFound = () => {
-  return (
-    <div className="min-h-screen flex items-center justify-center">
-      <h1 className="text-3xl font-bold">404 - Không tìm thấy trang</h1>
-    </div>
-  );
-};
-
-export default NotFound;
