+++ src/routes/index.jsx
@@ -1,24 +1,20 @@
-import { Spin } from 'antd';
 import { lazy, Suspense } from 'react';
-import { createBrowserRouter, Outlet } from 'react-router-dom';
-import LandingLayout from '../components/layouts/LandingLayout';
-import { PATH_NAME } from '../constants';
-import AuthPage from '../pages/auth';
-import NotFound from '../pages/notfound';
-import AdminRoutes from './AdminRoutes';
-import GuestRoute from './GuestRoute';
-import MemberRoutes from './MemberRoutes';
+import { createBrowserRouter } from 'react-router-dom';
 
-const LandingPage = lazy(() => import('../pages/landing'));
-const MemberPage = lazy(() => import('../pages/member'));
-const AdminHome = lazy(() => import('../pages/admin'));
-const Post = lazy(() => import('../pages/post'));
-const PostDetail = lazy(() => import('../pages/post/detail'));
+const LandingPage = lazy(() => import('../pages/LandingPage'));
+const LoginPage = lazy(() => import('../pages/auth/LoginPage'));
+const RegisterPage = lazy(() => import('../pages/auth/RegisterPage'));
+const MemberPage = lazy(() => import('../pages/member/MemberPage'));
+const AdminHome = lazy(() => import('../pages/admin/AdminHome'));
+
+import { Spin } from 'antd';
+import AdminRoutes from './AdminRoutes';
+import MemberRoutes from './MemberRoutes';
 
 const withSuspense = (Component) => (
   <Suspense
     fallback={
-      <div className="flex items-center justify-center min-h-screen">
+      <div>
         <Spin size="large" />
       </div>
     }
@@ -29,40 +25,22 @@ const withSuspense = (Component) => (
 
 const router = createBrowserRouter([
   {
-    element: (
-      <LandingLayout>
-        <Outlet />
-      </LandingLayout>
-    ),
-    children: [
-      {
-        path: PATH_NAME.HOME,
-        element: withSuspense(LandingPage),
-      },
-      {
-        path: PATH_NAME.POST,
-        element: withSuspense(Post),
-      },
-      {
-        path: PATH_NAME.POST_DETAIL,
-        element: withSuspense(PostDetail),
-      },
-    ],
+    path: '/',
+    element: withSuspense(LandingPage),
   },
   {
-    element: <GuestRoute />,
-    children: [
-      {
-        path: PATH_NAME.AUTH,
-        element: <AuthPage />,
-      },
-    ],
+    path: '/login',
+    element: withSuspense(LoginPage),
+  },
+  {
+    path: '/register',
+    element: withSuspense(RegisterPage),
   },
   {
     element: <MemberRoutes />,
     children: [
       {
-        path: PATH_NAME.MEMBER,
+        path: '/member',
         element: withSuspense(MemberPage),
       },
     ],
@@ -71,14 +49,10 @@ const router = createBrowserRouter([
     element: <AdminRoutes />,
     children: [
       {
-        path: PATH_NAME.ADMIN,
+        path: '/admin',
         element: withSuspense(AdminHome),
       },
     ],
-  },
-  {
-    path: PATH_NAME.NOT_FOUND,
-    element: <NotFound />,
   },
 ]);
